Updated code based on new thrift format

+++ core/src/main/java/org/carbondata/core/carbon/metadata/datatype/DataType.java
@@ -23,7 +23,7 @@ public enum DataType {
     STRING,
     INTEGER,
     LONG,
-    DECIMAL,
+    DOUBLE,
     BIG_DECIMAL,
     TIME_STAMP,
     ARRAY,
